+++ lib/stylish_dialog.dart
@@ -64,18 +64,6 @@ class StylishDialog {
   /// The widget can be any text-based widget, such as [Text] or [RichText].
   Widget? content;
 
-  ///Use this to set confirm button text.
-  ///
-  ///@Deprecated('Use `confirmEvent` instead. Will be removed soon')
-  @Deprecated('Use `confirmEvent` instead. Will be removed soon')
-  String? confirmText;
-
-  ///Use this to set cancel button text.
-  ///
-  ///@Deprecated('Use `cancelEvent` instead. Will be removed soon')
-  @Deprecated('Use `cancelEvent` instead. Will be removed soon')
-  String? cancelText;
-
   /// Use this property to specify whether the dialog should be dismissed when the user touches outside of it.
   /// If set to `true`, the dialog will be dismissed when the user touches outside of it.
   /// If set to `false`, the dialog will not be dismissed when the user touches outside of it.
@@ -94,19 +82,7 @@ class StylishDialog {
   ///   // other properties...
   /// )
   /// ```
-  bool dismissOnTouchOutside;
-
-  ///Hanlde confirm button press event.
-  ///
-  ///@Deprecated('Use `confirmButton` instead. Will be removed soon')
-  @Deprecated('Use `confirmButton` instead. Will be removed soon')
-  VoidCallback? confirmPressEvent;
-
-  ///Hanlde cancel button press event.
-  ///
-  ///@Deprecated('Use `cancelButton` instead. Will be removed soon')
-  @Deprecated('Use `cancelButton` instead. Will be removed soon')
-  VoidCallback? cancelPressEvent;
+  final bool dismissOnTouchOutside;
 
   /// Allows you to add a custom widget to the dialog.
   ///
@@ -120,7 +96,7 @@ class StylishDialog {
   ///   // other properties...
   /// )
   /// ```
-  Widget? addView;
+  final Widget? addView;
 
   /// Use this property to add a confirm button widget to the dialog.
   /// The widget can be any type of button, such as [TextButton], [ElevatedButton] or [IconButton].
@@ -220,7 +196,7 @@ class StylishDialog {
   ///   // other properties...
   /// )
   /// ```
-  Style? style;
+  final Style? style;
 
   /// Use this property to handle the callbacks of the dialog.
   ///
@@ -248,21 +224,13 @@ class StylishDialog {
   ///Default value is 48.0
   ///
   ///sets equal height and width to the progress Indicator Container
-  double? progressIndicatorSize;
+  final double? progressIndicatorSize;
 
   StylishDialog({
     required this.context,
     required this.alertType,
     this.title,
     this.content,
-    @Deprecated('Use `confirmButton` instead. will be removed soon')
-    this.confirmText,
-    @Deprecated('Use `cancelButton` instead. will be removed soon')
-    this.cancelText,
-    @Deprecated('Use `confirmButton` instead. will be removed soon')
-    this.confirmPressEvent,
-    @Deprecated('Use `cancelButton` instead. will be removed soon')
-    this.cancelPressEvent,
     this.dismissOnTouchOutside = true,
     this.addView,
     this.confirmButton,
@@ -288,8 +256,8 @@ class StylishDialog {
         builder: (context) {
           controller?.setValue = DialogStatus.Showing;
 
-          return StatefulBuilder(builder: (context, setState) {
-            _stateSetter = setState;
+          return StatefulBuilder(builder: (context, state) {
+            _stateSetter = state;
 
             return _buildDialog;
           });
@@ -304,10 +272,13 @@ class StylishDialog {
       });
 
   //to build dialog
-  Widget get _buildDialog =>
-      WillPopScope(onWillPop: _onWillPop, child: _buildDialogUI());
-
-  Future<bool> _onWillPop() async => dismissOnTouchOutside;
+  Widget get _buildDialog => PopScope(
+        canPop: dismissOnTouchOutside,
+        onPopInvoked: (onPop) {
+          _stateSetter = null;
+        },
+        child: _buildDialogUI(),
+      );
 
   ///The Function used to dismiss the dialog
   ///```
@@ -323,7 +294,7 @@ class StylishDialog {
     Navigator.of(context, rootNavigator: rootNavigator).pop();
   }
 
-  late StateSetter _stateSetter;
+  StateSetter? _stateSetter;
   StylishDialogType _changeAlert = StylishDialogType._CHANGE;
 
   ///Call [changeAlertType] function to change the alert type of the current dialog
@@ -348,14 +319,6 @@ class StylishDialog {
     /// The main content of the dialog, displayed in the middle of the dialog box.
     /// The widget can be any text-based widget, such as [Text] or [RichText].
     Widget? content,
-    @Deprecated('Use `confirmButton` instead. will be removed soon')
-    String? confirmText,
-    @Deprecated('Use `cancelButton` instead. will be removed soon')
-    String? cancelText,
-    @Deprecated('Use `confirmButton` instead. will be removed soon')
-    VoidCallback? confirmPressEvent,
-    @Deprecated('Use `cancelButton` instead. will be removed soon')
-    VoidCallback? cancelPressEvent,
 
     /// Use this property to add a confirm button widget to the dialog.
     /// The widget can be any type of button, such as [TextButton], [ElevatedButton] or [IconButton].
@@ -413,13 +376,10 @@ class StylishDialog {
     /// ```
     Widget? cancelButton,
   }) {
-    _stateSetter(() {
+    if (_stateSetter == null) return;
+    _stateSetter!(() {
       this.title = title;
       this.content = content;
-      this.confirmText = confirmText;
-      this.cancelText = cancelText;
-      this.confirmPressEvent = confirmPressEvent;
-      this.cancelPressEvent = cancelPressEvent;
       this.confirmButton = confirmButton;
       this.cancelButton = cancelButton;
 
@@ -439,10 +399,6 @@ class StylishDialog {
           _changeAlert == StylishDialogType._CHANGE ? alertType : _changeAlert,
       title: title,
       content: content,
-      confirmText: confirmText,
-      cancelText: cancelText,
-      confirmPressEvent: confirmPressEvent,
-      cancelPressEvent: cancelPressEvent,
       addView: addView,
       cancelButton: cancelButton,
       confirmButton: confirmButton,