+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 ## 1.0.2
+Changes: 
+- `Style.Default` to `DefaultStyle()`
+- `Style.Style1` to `Style1()`
+- 
 Fix: bug
+
 ## 1.0.1
 Fix: bug
+
 ## 1.0.1-beta
 Fix: bug
+
 ## 1.0.0
 ```diff
 + title: